requirements.txt
python-decouple==3.1
discord==1.7.3
discord.py==1.7.3
requests==2.26.0